move part of APIs from utils/CommonUtil to config/Utils
